Update todo list
I don't need the mailchimp module.
All the visitors are anonymous, so there's no need to integrate with
Mailchimp. It's enough to just have a sign up form embedded in the site.

I am not going to create a lead magnet right now.
I will see if I can get people to sign up just to be kept notified about
when the latest video has been added to the site.